double to float, update Q matrix with delta T

+++ include/tracker.h
@@ -11,7 +11,7 @@ public:
     void Predict(float dt);
     void Update(const cv::Rect& bbox);
     cv::Rect GetStateAsBbox() const;
-    double GetNIS() const;
+    float GetNIS() const;
 
     int coast_cycles_ = 0, hit_streak_ = 0;
